+++ src/main/java/com/example/shawarmahouse/model/MenuItem.java
@@ -1,7 +1,6 @@
 package com.example.shawarmahouse.model;
 
 import lombok.*;
-import org.bson.types.Binary;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -16,7 +15,7 @@ public class MenuItem {
     private String item_id;
     private String name;
     private String description;
-    private int price;
+    private long price;
     private String category;
     private String photo; // Storing the photo as binary data
     private String subCategory;
